third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/text-emphasis-color-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'text-emphasis-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
Harness: the test ran to completion.
